+++ src/index.ts
@@ -1,21 +1,6 @@
 import { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 import { StdioServerTransport } from "@modelcontextprotocol/sdk/server/stdio.js";
-import { z } from "zod";
-import dotenv from "dotenv";
-
-// Load environment variables from .env file
-dotenv.config();
-
-const ROLLBAR_API_BASE = "https://api.rollbar.com/api/1";
-const USER_AGENT = "rollbar-mcp-server/0.0.1";
-const ROLLBAR_ACCESS_TOKEN = process.env.ROLLBAR_ACCESS_TOKEN;
-
-if (!ROLLBAR_ACCESS_TOKEN) {
-  console.error(
-    "Error: ROLLBAR_ACCESS_TOKEN is not set in env var or .env file",
-  );
-  process.exit(1);
-}
+import { registerAllTools } from "./tools/index.js";
 
 // Create server instance
 const server = new McpServer({
@@ -39,326 +24,8 @@ const server = new McpServer({
   },
 });
 
-// Helper function for making Rollbar API requests
-async function makeRollbarRequest<T>(url: string): Promise<T | null> {
-  const headers = {
-    "User-Agent": USER_AGENT,
-    "X-Rollbar-Access-Token": ROLLBAR_ACCESS_TOKEN as string,
-    Accept: "application/json",
-  };
-
-  try {
-    const response = await fetch(url, { headers });
-    if (!response.ok) {
-      throw new Error(`HTTP error! status: ${response.status}`);
-    }
-    return (await response.json()) as T;
-  } catch (error) {
-    console.error("Error making Rollbar API request:", error);
-    return null;
-  }
-}
-
-// Define interfaces for Rollbar API responses
-interface RollbarApiResponse<T> {
-  err: number;
-  result: T;
-}
-
-interface RollbarItemResponse {
-  id: number;
-  counter: number;
-  environment: string;
-  framework: string;
-  title: string;
-  timestamp: number;
-  last_occurrence_id: number;
-  last_occurrence_timestamp: number;
-  level: string;
-  project_id: number;
-  language: string;
-  platform: string;
-  hash: string;
-  exception?: any;
-  request?: any;
-  body?: any;
-  data: {
-    body: any;
-    level: string;
-    environment: string;
-    framework: string;
-    language: string;
-    timestamp: number;
-    platform: string;
-    request?: {
-      url: string;
-      method: string;
-    };
-    exception?: {
-      class: string;
-      message: string;
-      description: string;
-    };
-    [key: string]: any; // Allow for any other properties
-  };
-  [key: string]: any; // Allow for any other properties
-}
-
-interface RollbarOccurrenceResponse {
-  id: number;
-  item_id: number;
-  timestamp: number;
-  version: number;
-  data: {
-    body: any;
-    level: string;
-    environment: string;
-    framework: string;
-    language: string;
-    timestamp: number;
-    platform: string;
-    request?: {
-      url: string;
-      method: string;
-    };
-    exception?: {
-      class: string;
-      message: string;
-      description: string;
-    };
-    context?: string;
-    code_version?: string;
-    stack_trace?: any;
-    metadata?: any;
-    [key: string]: any; // Allow for any other properties
-  };
-  [key: string]: any; // Allow for any other properties
-}
-
-interface RollbarDeployResponse {
-  environment: string;
-  revision: string;
-  local_username: string;
-  comment: string;
-  status: string;
-  id: number;
-  project_id: number;
-  user_id: number;
-  start_time: number;
-  finish_time: number;
-  [key: string]: any; // Allow for any other properties
-}
-
-interface RollbarVersionsResponse {
-  id: number;
-  version: string;
-  environment: string;
-  date_created: number;
-  first_occurrence_id: number;
-  first_occurrence_timestamp: number;
-  last_occurrence_id: number;
-  last_occurrence_timestamp: number;
-  deployed_by: string;
-  last_deploy_timestamp: number;
-  [key: string]: any; // Allow for any other properties
-}
-
-// Register tools
-server.tool(
-  "get-item-details",
-  "Get item details for a Rollbar item",
-  {
-    counter: z.number().int().describe("Rollbar item counter"),
-  },
-  async ({ counter }) => {
-    try {
-      // Redirects are followed, so we get an item response from the counter request
-      const counterUrl = `${ROLLBAR_API_BASE}/item_by_counter/${counter}`;
-      const itemResponse =
-        await makeRollbarRequest<RollbarApiResponse<RollbarItemResponse>>(
-          counterUrl,
-        );
-      console.error(itemResponse);
-
-      if (!itemResponse || itemResponse.err !== 0) {
-        return {
-          content: [
-            {
-              type: "text",
-              text: `Failed to retrieve item details for counter ${counter}.`,
-            },
-          ],
-        };
-      }
-
-      const item = itemResponse.result;
-
-      // Use the complete item data
-      const formattedData = item;
-
-      const occurrenceUrl = `${ROLLBAR_API_BASE}/instance/${item.last_occurrence_id}`;
-      console.error(`Fetching occurrence details from: ${occurrenceUrl}`);
-      const occurrenceResponse =
-        await makeRollbarRequest<RollbarApiResponse<RollbarOccurrenceResponse>>(
-          occurrenceUrl,
-        );
-      console.error("Occurrence response:", occurrenceResponse);
-
-      if (!occurrenceResponse || occurrenceResponse.err !== 0) {
-        // We got the item but failed to get occurrence, return just the item data
-        return {
-          content: [
-            {
-              type: "text",
-              text: JSON.stringify(formattedData, null, 2),
-            },
-          ],
-        };
-      }
-
-      const occurrence = occurrenceResponse.result;
-
-      // Remove the metadata section from occurrence.data to avoid exposing sensitive information
-      if (occurrence.data && occurrence.data.metadata) {
-        delete occurrence.data.metadata;
-      }
-
-      // Combine item and occurrence data
-      const responseData = {
-        ...formattedData,
-        occurrence: occurrence,
-      };
-
-      return {
-        content: [
-          {
-            type: "text",
-            text: JSON.stringify(responseData, null, 2),
-          },
-        ],
-      };
-    } catch (error) {
-      console.error("Error in get-item-details tool:", error);
-      return {
-        content: [
-          {
-            type: "text",
-            text: `Error retrieving item details: ${error instanceof Error ? error.message : String(error)}`,
-          },
-        ],
-      };
-    }
-  },
-);
-
-server.tool(
-  "get-deployments",
-  "Get deployments data from Rollbar",
-  {
-    limit: z
-      .number()
-      .int()
-      .describe("Number of Rollbar deployments to retrieve"),
-  },
-  async ({ limit }) => {
-    try {
-      const deploysUrl = `${ROLLBAR_API_BASE}/deploys?limit=${limit}`;
-      const deploysResponse =
-        await makeRollbarRequest<RollbarApiResponse<RollbarDeployResponse>>(
-          deploysUrl,
-        );
-      console.error(deploysResponse);
-
-      if (!deploysResponse || deploysResponse.err !== 0) {
-        return {
-          content: [
-            {
-              type: "text",
-              text: `Failed to retrieve deployments.`,
-            },
-          ],
-        };
-      }
-
-      const deployItem = deploysResponse.result;
-      console.error("Deployments response:", deployItem);
-
-      return {
-        content: [
-          {
-            type: "text",
-            text: JSON.stringify(deployItem, null, 2),
-          },
-        ],
-      };
-    } catch (error) {
-      console.error("Error in get-deployments tool:", error);
-      return {
-        content: [
-          {
-            type: "text",
-            text: `Error retrieving deployment details: ${error instanceof Error ? error.message : String(error)}`,
-          },
-        ],
-      };
-    }
-  },
-);
-
-server.tool(
-  "get-version",
-  "Get version details for a Rollbar project",
-  {
-    version: z.coerce.string().describe("GitSHA version to filter by"),
-    environment: z.coerce
-      .string()
-      .default("production")
-      .describe("Environment to filter by (default: production)"),
-  },
-  async ({ version, environment }) => {
-    try {
-      const versionsUrl = `${ROLLBAR_API_BASE}/versions/${version}?environment=${environment}`;
-      const versionsResponse =
-        await makeRollbarRequest<RollbarApiResponse<RollbarVersionsResponse>>(
-          versionsUrl,
-        );
-      console.error(versionsResponse);
-
-      if (!versionsResponse || versionsResponse.err !== 0) {
-        return {
-          content: [
-            {
-              type: "text",
-              text: `Failed to retrieve version data.`,
-            },
-          ],
-        };
-      }
-
-      const versionItem = versionsResponse.result;
-      console.error("Versions response:", versionItem);
-
-      return {
-        content: [
-          {
-            type: "text",
-            text: JSON.stringify(versionItem, null, 2),
-          },
-        ],
-      };
-    } catch (error) {
-      console.error("Error in get-versions tool:", error);
-      return {
-        content: [
-          {
-            type: "text",
-            text: `Error retrieving versions details: ${error instanceof Error ? error.message : String(error)}`,
-          },
-        ],
-      };
-    }
-  },
-);
+// Register all tools
+registerAllTools(server);
 
 async function main() {
   const transport = new StdioServerTransport();
